refactor: object_commands.py 클래스별 파일 분리 + i18n 전환 (태스크 25.3)
- object_commands.py(1190줄) → 7개 파일로 분리
  - get_command.py, drop_command.py, inventory_command.py
  - examine_command.py, equip_command.py, use_command.py
  - object_commands.py는 re-export 파일로 유지 (하위 호환성)
- 하드코딩 한국어 메시지 약 40곳 → I18N.get_message() 전환
- item.json에 obj.* i18n 키 약 20개 추가
- design.md에 향후 아키텍처 방향 (GUI 클라이언트 + i18n 전환 계획) 추가
- GetCommand dead code 제거
