+++ .kiro/specs/python-mud-engine/design.md
@@ -811,6 +811,19 @@ NPC:
 - 세션 타임아웃 (300초 비활성)
 - 로그인 시도 횟수 제한 (3회)
 
+## 향후 아키텍처 방향
+
+### GUI 클라이언트 및 i18n 전환 계획
+
+장기적으로 GUI 클라이언트를 개발할 예정이며, i18n(다국어 처리)은 클라이언트 측에서 담당하게 됩니다.
+
+- 현재: 서버에서 LocalizationManager를 통해 locale별 메시지를 생성하여 Telnet 클라이언트에 전송
+- 향후: 서버는 언어 중립적인 메시지 키 + 파라미터를 전송하고, GUI 클라이언트가 로컬에서 번역 처리
+- 전환 시: 서버의 `I18N.get_message()` 호출을 메시지 키 + 데이터 구조 전송으로 교체
+- Telnet 클라이언트는 레거시로 유지하되, 서버 측 i18n은 Telnet 호환 레이어로 분리
+
+이 전환이 이루어지기 전까지는 서버 측 LocalizationManager 기반 i18n을 유지합니다.
+
 ## 의존성
 
 - Python 3.13
